Ielts _pages
IELTS Academic and IELTS General Training are designed to cover the full range of abilities from non-user to expert user. The Academic version is for test takers who want to study at the tertiary level in an English-speaking country or seek professional registration. The General Training version is for test takers who want to work, train, study at a secondary school or migrate to an English-speaking country.